+++ packages/kit/src/routes/config/allowList.ts
@@ -2,7 +2,7 @@ import platformEnv from '@onekeyhq/shared/src/platformEnv';
 
 import { ERootRoutes } from '../enum';
 import { EGalleryRoutes } from '../Tab/Developer/Gallery/routes';
-import { ETabHomeRoutes } from '../Tab/Home/Routes';
+import { ETabHomeRoutes } from '../Tab/Home/router';
 import { ETabSwapRoutes } from '../Tab/Swap/Routes';
 import { ETabRoutes } from '../Tab/type';
