+++ packages/kit/src/views/TestModal/pages/TestSimpleModal.tsx
@@ -14,7 +14,7 @@ import HeaderIconButton from '@onekeyhq/components/src/layouts/Navigation/Header
 import useAppNavigation from '../../../hooks/useAppNavigation';
 import { ETestModalPages } from '../router/type';
 
-import type { ITabHomeParamList } from '../../../routes/Tab/Home/Routes';
+import type { ITabHomeParamList } from '../../../routes/Tab/Home/router';
 
 export function TestSimpleModal() {
   const headerRightCall = useCallback(
