+++ packages/kit/src/views/iOSFullScreenTestModal/pages/TestIOSFullModal.tsx
@@ -7,7 +7,7 @@ import HeaderIconButton from '@onekeyhq/components/src/layouts/Navigation/Header
 import useAppNavigation from '../../../hooks/useAppNavigation';
 import { EIOSFullScreenModalRoutes } from '../../../routes/iOSFullScreen/type';
 
-import type { ITabHomeParamList } from '../../../routes/Tab/Home/Routes';
+import type { ITabHomeParamList } from '../../../routes/Tab/Home/router';
 
 export function TestIOSFullModal() {
   const headerRightCall = useCallback(
